% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Areal_data.R
\name{Areal_data}
\alias{Areal_data}
\title{Main function to generate spatial features using areal data}
\usage{
Areal_data(
  path = "/Users/lix233/Haystack/5862_cell_centers/",
  file = "f8008cf4-3fe7-4200-a2af-21c27708ae28.csv",
  cell_class_var = "cell_class",
  x_var = "X0",
  y_var = "X1",
  from_type = "Lymphocyte",
  to_type = "Tumor",
  scale,
  myplot = FALSE
)
}
\arguments{
\item{path}{The path for the directory that contains the data}

\item{file}{The file name in the directory}

\item{cell_class_var}{The column name in each file that indicates the cell class variable}

\item{x_var}{The column name in each file that indicates the cell location x variable}

\item{y_var}{The column name in each file that indicates the cell location y variable}

\item{from_type}{The cell type one wants to use as the "from" cell type}

\item{to_type}{The cell type one wants to use as the "to" cell type}

\item{scale}{The spatial range that user wants to investigate}

\item{myplot}{Whether to plot the results, if available, by default it is set as FALSE}
}
\value{
This function returns the features for areal data (denote from_type as type 1 and to_type as type 2), using spatial lattice process methods:
\item{BC}{Bhattacharyya coefficient of type 1&2}
\item{MH_index}{Morisita-Horn index of type 1&2}
\item{JaccardJ}{Jaccard index of type 1&2}                                 
\item{SorensenL}{Sorensen index of type 1&2}
\item{Moran_I_tumor}{global Moran's I of type 1}
\item{Moran_I_immune}{global Moran's I of type 2}
\item{moran_I_Bivariate}{global bivariate Moran's I of type 1&2}
\item{geary_TC}{global Geary's C of type 1}
\item{geary_IC}{global Geary's C of type 2}
\item{moran_HL_TC}{percentage of significant high-low local Moran's I of type 1}
\item{moran_HH_TC}{percentage of significant high-high local Moran's I of type 1}
\item{moran_LH_TC}{percentage of significant low-high local Moran's I of type 1}
\item{moran_LL_TC}{percentage of significant low-low local Moran's I of type 1}
\item{geary_HH_TC}{percentage of significant high-high local Geary's C of type 1}
\item{geary_LL_TC}{percentage of significant low-low local Geary's C of type 1}
\item{moran_HL_IC}{percentage of significant high-low local Moran's I of type 2}
\item{moran_HH_IC}{percentage of significant high-high local Moran's I of type 2}
\item{moran_LH_IC}{percentage of significant low-high local Moran's I of type 2}
\item{moran_LL_IC}{percentage of significant low-low local Moran's I of type 2}
\item{geary_HH_IC}{percentage of significant high-high local Geary's C of type 2}
\item{geary_LL_IC}{percentage of significant low-low local Geary's C of type 2}
\item{GetisOrd_HS}{percentage of significant region being hotspot for both type 1 and type 2}
\item{GetisOrd_CS}{percentage of significant region being coldspot for both type 1 and type 2}
\item{GetisOrd_CS_IC_HS_TC}{percentage of significant region being hotspot for type 1 and coldspot for type 2}
\item{GetisOrd_CS_TC_HS_IC}{percentage of significant region being coldspot for type 1 and hotspot for type 2}
\item{GetisOrd_HS_IC}{percentage of significant region being hotspot for type 2}
\item{GetisOrd_CS_IC}{percentage of significant region being coldspot for type 2}
\item{GetisOrd_HS_TC}{percentage of significant region being hotspot for type 1}
\item{GetisOrd_CS_TC}{percentage of significant region being coldspot for type 1}
\item{GetisOrd_S_intra_cancer}{GetisOrd_HS/(GetisOrd_HS + GetisOrd_CS_IC_HS_TC + 0.000001), 0.000001 added to avoid denominator being 0}
\item{GetisOrd_S_intra_immune}{GetisOrd_HS/(GetisOrd_HS + GetisOrd_CS_TC_HS_IC + 0.000001), 0.000001 added to avoid denominator being 0}
\item{Lee_L}{global Lee's L statistics of type 1 and type 2}
\item{Lee_HL_TC_IC}{percentage of significant high-low local Lee's L from type 1 to type 2}
\item{Lee_HH_TC_IC}{percentage of significant high-high local Lee's L from type 1 to type 2}
\item{Lee_LH_TC_IC}{percentage of significant low-high local Lee's L from type 1 to type 2}
\item{Lee_LL_TC_IC}{percentage of significant low-low local Lee's L from type 1 to type 2}
\item{Lee_HL_IC_TC}{percentage of significant high-low local Lee's L from type 2 to type 1}
\item{Lee_HH_IC_TC}{percentage of significant high-high local Lee's L from type 2 to type 1}
\item{Lee_LH_IC_TC}{percentage of significant low-high local Lee's L from type 2 to type 1}
\item{Lee_LL_IC_TC}{percentage of significant low-low local Lee's L from type 2 to type 1}
}
\description{
This is a function that calculates the features for lattice data, using spatial lattice process.
Methods included in this function include global and local autocorrelation indices, e.g. Moran's I, Geary's C, Getis-Ord statistics.
Co-localization indices are also calculated, such as Morisita-Horn index, Sorensen index, and Jaccard index
}
\author{
Xiao Li, \email{xiao.li.xl2@roche.com}
}
