% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Vis.R
\name{Data_Vis}
\alias{Data_Vis}
\title{A Data_Vis function}
\usage{
Data_Vis(
  path = "/Users/lix233/Haystack/5862_cell_centers/",
  file = "f8008cf4-3fe7-4200-a2af-21c27708ae28.csv",
  cell_class_var = "cell_class",
  x_var = "X0",
  y_var = "X1",
  cell_type = "Lymphocyte"
)
}
\arguments{
\item{path}{The path for the directory that contains the data}

\item{file}{The file name in the directory}

\item{cell_class_var}{The column name in each file that indicates the cell class variable}

\item{x_var}{The column name in each file that indicates the cell location x variable}

\item{y_var}{The column name in each file that indicates the cell location y variable}

\item{cell_type}{The cell type one wants to use as the univariate marker}
}
\value{
This function returns spatial distribution plot for certain cell type
}
\description{
This is a function that visualizes the input cell level data
}
\author{
Xiao Li, \email{li.xiao@gene.com}
}
