% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delaunay_triangulation_cross.R
\name{Delaunay_triangulation_cross}
\alias{Delaunay_triangulation_cross}
\title{A cross type Delaunay triangulation function with comprehensive feature extraction}
\usage{
Delaunay_triangulation_cross(spp_df, cell_type, scale, myplot = FALSE)
}
\arguments{
\item{spp_df}{The input spatial data.frame, needs columns: "x", "y", "cell_class".}

\item{cell_type}{The cell types that graph features are to be calculated for.}

\item{scale}{The threshold to trim long graph edges.}

\item{myplot}{Whether to plot the results (default: FALSE).}
}
\value{
A named list containing a comprehensive set of graph features.
(See documentation for DT_graph_cross_subregion_random for a full list).
}
\description{
This function builds a Delaunay graph and calculates a rich set of features describing
the spatial architecture, including global, node-level, and neighborhood metrics.
}
\author{
Xiao Li, \email{xiao.li.xl2@roche.com}
}
