% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_summary.R
\name{distribution_summary}
\alias{distribution_summary}
\title{distribution_summary}
\usage{
distribution_summary(x)
}
\arguments{
\item{x}{The input vector}
}
\value{
This function returns the summary features from a distribution
}
\description{
This is the function that calculates the summary features from a distribution
}
\author{
Xiao Li, \email{xiao.li.xl2@roche.com}
}
