% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_graph_cross_subregion_random.R
\name{DT_graph_cross_subregion_random}
\alias{DT_graph_cross_subregion_random}
\title{The main function that calculates a rich set of Delaunay triangulation graph features.}
\usage{
DT_graph_cross_subregion_random(
  path,
  file,
  cell_type = c("tumor", "lymphocyte"),
  cell_class_var = "cell_class",
  x_var = "X0",
  y_var = "X1",
  scale = 1000,
  side_length = 2000,
  num_FOV = 5,
  set_seed = 42,
  myplot = FALSE
)
}
\arguments{
\item{path}{The path for the directory that contains the data}

\item{file}{The file name in the directory}

\item{cell_type}{The cell types that graph features are to be calculated}

\item{cell_class_var}{The column name in each file that indicates the cell class variable}

\item{x_var}{The column name in each file that indicates the cell location x variable}

\item{y_var}{The column name in each file that indicates the cell location y variable}

\item{scale}{The threshold to trim the graph edges. This is especially useful when the edge connects cells that are from different tissue region, e.g. different tumor nests. If user wants to retain all the edges, set it to a very large number.}

\item{side_length}{The side length of squared FOVs}

\item{num_FOV}{Number of FOVs to choose}

\item{set_seed}{set seed for reproducibility}

\item{myplot}{Whether to plot the results, if available, by default it is set as FALSE}
}
\value{
This function returns a data frame with the median value of each feature across the sampled FOVs:
  \item{--- Global Graph Features ---}{}
  \item{graph_assortativity}{A score from -1 to 1. Positive values indicate spatial segregation (cells prefer same-type neighbors). Negative values indicate cell mixing.}
  \item{num_homotypic_edges}{The total count of edges connecting cells of the same type.}
  \item{num_heterotypic_edges}{The total count of edges connecting cells of different types.}
  \item{--- Node-Level Features (Aggregated per cell type) ---}{}
  \item{mean_degree_[cell_type]}{Average number of direct neighbors for a given cell type.}
  \item{mean_lcc_[cell_type]}{Average local clustering coefficient for a given cell type, measuring neighborhood "cliquishness".}
  \item{--- Neighborhood Composition Features (Interaction Quantification) ---}{}
  \item{avg_prop_[type_A]_neighbor_for_[type_B]}{For an average cell of type B, this is the proportion of its neighbors that are of type A.}
  \item{--- Original Edge Weight & Infiltration Features ---}{}
  \item{infiltration_score_s_k}{Ratio of heterotypic (s-k) to homotypic (s-s) edge counts.}
  \item{mean_x_cross_weight}{Average length of edges connecting different cell types.}
  \item{... (and all other edge weight statistics from your original script) ...}{}

#' @import magrittr
}
\description{
This function calculates Delaunay triangulation graph features for distinct cell types
based on randomly sampled Fields of View (FOVs). It returns a comprehensive set of
features describing the tissue architecture at global, local, and cellular levels.
}
\author{
Xiao Li, \email{xiao.li.xl2@roche.com}
}
