% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_graph_uni_subregion_random.R
\name{DT_graph_uni_subregion_random}
\alias{DT_graph_uni_subregion_random}
\title{The main function that calculate the Delaunay_triangulation graph features for a single cell type}
\usage{
DT_graph_uni_subregion_random(
  path,
  file,
  cell_type = "lymphocyte",
  cell_class_var = "cell_class",
  x_var = "X0",
  y_var = "X1",
  scale = 1000,
  side_length = 2000,
  num_FOV = 5,
  set_seed = 42,
  myplot = FALSE
)
}
\arguments{
\item{path}{The path for the directory that contains the data}

\item{file}{The file name in the directory}

\item{cell_type}{The cell type that graph features are to be calculated}

\item{cell_class_var}{The column name in each file that indicates the cell class variable}

\item{x_var}{The column name in each file that indicates the cell location x variable}

\item{y_var}{The column name in each file that indicates the cell location y variable}

\item{scale}{The threshold to trim the graph edges. This is especially useful when the edge connects cells that are from different tissue region, e.g. different tumor nests. If user wants to retain all the edges, set it to a very large number.}

\item{side_length}{The side length of squared FOVs}

\item{num_FOV}{Number of FOVs to choose}

\item{set_seed}{set seed for reproducibility}

\item{myplot}{Whether to plot the results, if available, by default it is set as FALSE}
}
\value{
This function returns the median value of each each feature across the FOVs:
\item{mean_x_weight}{average edge length of the graph}
\item{median_x_weight}{median edge length of the graph}
\item{sd_x_weight}{standard deviation of edge length of the graph}
\item{iqr_x_weight}{IQR of edge length of the graph}
\item{skewness_x_weight}{skewness of edge length of the graph}
\item{kurtosis_x_weight}{kurtosis of edge length of the graph}
\item{min_x_weight}{minimum edge length of the graph}
\item{max_x_weight}{maximum edge length of the graph}
\item{range_x_weight}{range of edge length of the graph}
\item{Q10_weight}{10\% quantile edge length of the graph}
\item{Q20_weight}{20\% quantile edge length of the graph}
\item{Q25_weight}{25\% quantile edge length of the graph}
\item{Q30_weight}{30\% quantile edge length of the graph}
\item{Q40_weight}{40\% quantile edge length of the graph}
\item{Q60_weight}{60\% quantile edge length of the graph}
\item{Q70_weight}{70\% quantile edge length of the graph}
\item{Q75_weight}{75\% quantile edge length of the graph}
\item{Q80_weight}{80\% quantile edge length of the graph}
\item{Q90_weight}{90\% quantile edge length of the graph}
\item{theil_index_weight}{theil index edge length of the graph}
\item{gini_coeff_weight}{gini index edge length of the graph}
\item{mean_x_closeness}{average node closeness centrality of the graph}
\item{median_x_closeness}{median node closeness centrality of the graph}
\item{sd_x_closeness}{standard deviation of node closeness centrality of the graph}
\item{iqr_x_closeness}{IQR of node closeness centrality of the graph}
\item{skewness_x_closeness}{skewness of node closeness centrality of the graph}
\item{kurtosis_x_closeness}{kurtosis of node closeness centrality of the graph}
\item{min_x_closeness}{minimum node closeness centrality of the graph}
\item{max_x_closeness}{maximum node closeness centrality of the graph}
\item{range_x_closeness}{range of node closeness centrality of the graph}
\item{Q10_closeness}{10\% quantile node closeness centrality of the graph}
\item{Q20_closeness}{20\% quantile node closeness centrality of the graph}
\item{Q25_closeness}{25\% quantile node closeness centrality of the graph}
\item{Q30_closeness}{30\% quantile node closeness centrality of the graph}
\item{Q40_closeness}{40\% quantile node closeness centrality of the graph}
\item{Q60_closeness}{60\% quantile node closeness centrality of the graph}
\item{Q70_closeness}{70\% quantile node closeness centrality of the graph}
\item{Q75_closeness}{75\% quantile node closeness centrality of the graph}
\item{Q80_closeness}{80\% quantile node closeness centrality of the graph}
\item{Q90_closeness}{90\% quantile node closeness centrality of the graph}
\item{theil_index_closeness}{theil index node closeness centrality of the graph}
\item{gini_coeff_closeness}{gini index node closeness centrality of the graph}
\item{mean_x_betweenness}{average node betweenness centrality of the graph}
\item{median_x_betweenness}{median node betweenness centrality of the graph}
\item{sd_x_betweenness}{standard deviation of node betweenness centrality of the graph}
\item{iqr_x_betweenness}{IQR of node betweenness centrality of the graph}
\item{skewness_x_betweenness}{skewness of node betweenness centrality of the graph}
\item{kurtosis_x_betweenness}{kurtosis of node betweenness centrality of the graph}
\item{min_x_betweenness}{minimum node betweenness centrality of the graph}
\item{max_x_betweenness}{maximum node betweenness centrality of the graph}
\item{range_x_betweenness}{range of node betweenness centrality of the graph}
\item{Q10_betweenness}{10\% quantile node betweenness centrality of the graph}
\item{Q20_betweenness}{20\% quantile node betweenness centrality of the graph}
\item{Q25_betweenness}{25\% quantile node betweenness centrality of the graph}
\item{Q30_betweenness}{30\% quantile node betweenness centrality of the graph}
\item{Q40_betweenness}{40\% quantile node betweenness centrality of the graph}
\item{Q60_betweenness}{60\% quantile node betweenness centrality of the graph}
\item{Q70_betweenness}{70\% quantile node betweenness centrality of the graph}
\item{Q75_betweenness}{75\% quantile node betweenness centrality of the graph}
\item{Q80_betweenness}{80\% quantile node betweenness centrality of the graph}
\item{Q90_betweenness}{90\% quantile node betweenness centrality of the graph}
\item{theil_index_betweenness}{theil index node betweenness centrality of the graph}
\item{gini_coeff_betweenness}{gini index node betweenness centrality of the graph}
\item{mean_x_degree}{average node degree centrality of the graph}
\item{median_x_degree}{median node degree centrality of the graph}
\item{sd_x_degree}{standard deviation of node degree centrality of the graph}
\item{iqr_x_degree}{IQR of node degree centrality of the graph}
\item{skewness_x_degree}{skewness of node degree centrality of the graph}
\item{kurtosis_x_degree}{kurtosis of node degree centrality of the graph}
\item{min_x_degree}{minimum node degree centrality of the graph}
\item{max_x_degree}{maximum node degree centrality of the graph}
\item{range_x_degree}{range of node degree centrality of the graph}
\item{Q10_degree}{10\% quantile node degree centrality of the graph}
\item{Q20_degree}{20\% quantile node degree centrality of the graph}
\item{Q25_degree}{25\% quantile node degree centrality of the graph}
\item{Q30_degree}{30\% quantile node degree centrality of the graph}
\item{Q40_degree}{40\% quantile node degree centrality of the graph}
\item{Q60_degree}{60\% quantile node degree centrality of the graph}
\item{Q70_degree}{70\% quantile node degree centrality of the graph}
\item{Q75_degree}{75\% quantile node degree centrality of the graph}
\item{Q80_degree}{80\% quantile node degree centrality of the graph}
\item{Q90_degree}{90\% quantile node degree centrality of the graph}
\item{theil_index_degree}{theil index node degree centrality of the graph}
\item{gini_coeff_degree}{gini index node degree centrality of the graph}
\item{mean_x_triangle_area}{average Delaunay triangle area of the graph}
\item{median_x_triangle_area}{median Delaunay triangle area of the graph}
\item{sd_x_triangle_area}{standard deviation of Delaunay triangle area of the graph}
\item{iqr_x_triangle_area}{IQR of Delaunay triangle area of the graph}
\item{skewness_x_triangle_area}{skewness of Delaunay triangle area of the graph}
\item{kurtosis_x_triangle_area}{kurtosis of Delaunay triangle area of the graph}
\item{min_x_triangle_area}{minimum Delaunay triangle area of the graph}
\item{max_x_triangle_area}{maximum Delaunay triangle area of the graph}
\item{range_x_triangle_area}{range of Delaunay triangle area of the graph}
\item{Q10_triangle_area}{10\% quantile Delaunay triangle area of the graph}
\item{Q20_triangle_area}{20\% quantile Delaunay triangle area of the graph}
\item{Q25_triangle_area}{25\% quantile Delaunay triangle area of the graph}
\item{Q30_triangle_area}{30\% quantile Delaunay triangle area of the graph}
\item{Q40_triangle_area}{40\% quantile Delaunay triangle area of the graph}
\item{Q60_triangle_area}{60\% quantile Delaunay triangle area of the graph}
\item{Q70_triangle_area}{70\% quantile Delaunay triangle area of the graph}
\item{Q75_triangle_area}{75\% quantile Delaunay triangle area of the graph}
\item{Q80_triangle_area}{80\% quantile Delaunay triangle area of the graph}
\item{Q90_triangle_area}{90\% quantile Delaunay triangle area of the graph}
\item{theil_index_triangle_area}{theil index Delaunay triangle area of the graph}
\item{gini_coeff_triangle_area}{gini index Delaunay triangle area of the graph}
\item{mean_x_triangle_perimeter}{average Delaunay triangle perimeter of the graph}
\item{median_x_triangle_perimeter}{median Delaunay triangle perimeter of the graph}
\item{sd_x_triangle_perimeter}{standard deviation of Delaunay triangle perimeter of the graph}
\item{iqr_x_triangle_perimeter}{IQR of Delaunay triangle perimeter of the graph}
\item{skewness_x_triangle_perimeter}{skewness of Delaunay triangle perimeter of the graph}
\item{kurtosis_x_triangle_perimeter}{kurtosis of Delaunay triangle perimeter of the graph}
\item{min_x_triangle_perimeter}{minimum Delaunay triangle perimeter of the graph}
\item{max_x_triangle_perimeter}{maximum Delaunay triangle perimeter of the graph}
\item{range_x_triangle_perimeter}{range of Delaunay triangle perimeter of the graph}
\item{Q10_triangle_perimeter}{10\% quantile Delaunay triangle perimeter of the graph}
\item{Q20_triangle_perimeter}{20\% quantile Delaunay triangle perimeter of the graph}
\item{Q25_triangle_perimeter}{25\% quantile Delaunay triangle perimeter of the graph}
\item{Q30_triangle_perimeter}{30\% quantile Delaunay triangle perimeter of the graph}
\item{Q40_triangle_perimeter}{40\% quantile Delaunay triangle perimeter of the graph}
\item{Q60_triangle_perimeter}{60\% quantile Delaunay triangle perimeter of the graph}
\item{Q70_triangle_perimeter}{70\% quantile Delaunay triangle perimeter of the graph}
\item{Q75_triangle_perimeter}{75\% quantile Delaunay triangle perimeter of the graph}
\item{Q80_triangle_perimeter}{80\% quantile Delaunay triangle perimeter of the graph}
\item{Q90_triangle_perimeter}{90\% quantile Delaunay triangle perimeter of the graph}
\item{theil_index_triangle_perimeter}{theil index Delaunay triangle perimeter of the graph}
\item{gini_coeff_triangle_perimeter}{gini index Delaunay triangle perimeter of the graph}
}
\description{
This is a function that calculate the Delaunay_triangulation graph features for a single cell type, based on random FOVs. The FOV size and number of FOVs are decided by decided by the user.
}
\author{
Xiao Li, \email{xiao.li.xl2@roche.com}
}
