% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardJ.R
\name{JaccardJ}
\alias{JaccardJ}
\title{Jaccard's J index}
\usage{
JaccardJ(x, y)
}
\arguments{
\item{x}{The coordinate x}

\item{y}{The coordinate y}
}
\value{
This function returns the Jaccard's J index
}
\description{
This is the function that calculates the Jaccard's J index
}
\author{
Xiao Li, \email{li.xiao@gene.com}
}
