% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_entropy.R
\name{local_entropy}
\alias{local_entropy}
\title{Main function to generate local entropy for each single cell}
\usage{
local_entropy(
  path = "/Users/lix233/Haystack/5813_cell_centers/",
  file = "267a5ed2-b6ee-4973-bf2d-3526b668eda0.csv",
  cell_class_var = "cell_class",
  x_var = "X0",
  y_var = "X1",
  scale = 200,
  myplot = T
)
}
\arguments{
\item{path}{The path for the directory that contains the data}

\item{file}{The file name in the directory}

\item{cell_class_var}{The column name in each file that indicates the cell class variable}

\item{x_var}{The column name in each file that indicates the cell location x variable}

\item{y_var}{The column name in each file that indicates the cell location y variable}

\item{scale}{The spatial range that user wants to define as local neighborhood}

\item{myplot}{Whether to plot the results, if available, by default it is set as FALSE}
}
\value{
This function returns a vector of local entropy for each single cell 
\item{local_entropy}{a vector of local entropy for each single cell}
}
\description{
This is a function that calculates the local entropy for each single cell
}
\author{
Xiao Li, \email{xiao.li.xl2@roche.com}
}
