% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Morisita_Horn.R
\name{Morisita_Horn}
\alias{Morisita_Horn}
\title{Morisita_Horn index}
\usage{
Morisita_Horn(x, y)
}
\arguments{
\item{x}{The coordinate x}

\item{y}{The coordinate y}
}
\value{
This function returns the Morisita_Horn index
}
\description{
This is the function that calculates the Morisita_Horn index
}
\author{
Xiao Li, \email{li.xiao@gene.com}
}
