% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_coords.R
\name{normalized_coords}
\alias{normalized_coords}
\title{Normalizes the coordinates to 0-1 scale}
\usage{
normalized_coords(x, L)
}
\arguments{
\item{x}{The original coordinate x or y, before normalization}

\item{L}{The scaling factor for normalization}
}
\value{
This function returns the normalized coordinates
}
\description{
This is a simple function that normalizes the coordinates to 0-1 scale
}
\author{
Xiao Li, \email{li.xiao@gene.com}
}
