% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variogram.R
\name{plot_variogram}
\alias{plot_variogram}
\title{Plot the semi-variogram/cross-variogram for geostatistics data}
\usage{
plot_variogram(v, m, title)
}
\arguments{
\item{v}{The sample variogram}

\item{m}{The fitted variogram}

\item{title}{The title appears in the plot}
}
\value{
This function returns semi-variogram/cross-variogram plot for geostatistics data
}
\description{
This is a function that plot the semi-variogram/cross-variogram for geostatistics data
}
\author{
Xiao Li, \email{xiao.li.xl2@roche.com}
}
