% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point_pattern_data_ITLR.R
\name{Point_pattern_data_ITLR}
\alias{Point_pattern_data_ITLR}
\title{Discrimination of target cell population}
\usage{
Point_pattern_data_ITLR(
  path = "/Users/lix233/Haystack/5862_cell_centers/",
  file = "f8008cf4-3fe7-4200-a2af-21c27708ae28.csv",
  cell_class_var = "cell_class",
  x_var = "X0",
  y_var = "X1",
  from_type = "Lymphocyte",
  to_type = "Tumor",
  micron_per_pixel = 1,
  myplot = FALSE
)
}
\arguments{
\item{path}{The path for the directory that contains the data}

\item{file}{The file name in the directory}

\item{cell_class_var}{The column name in each file that indicates the cell class variable}

\item{x_var}{The column name in each file that indicates the cell location x variable}

\item{y_var}{The column name in each file that indicates the cell location y variable}

\item{from_type}{The cell type one wants to use as the "from" cell type}

\item{to_type}{The cell type one wants to use as the "to" cell type}

\item{micron_per_pixel}{The ratio of pixel to micron, i.e. scan resolution}

\item{myplot}{Whether to plot the results, if available, by default it is set as FALSE}
}
\value{
This function discriminate a target cell population into 2 subgroups, denote them as ITL and ATL. 

\item{ITLR}{ratio of number of ITL versus number of (ITL + ATL)}
\item{ITLR2}{ratio of number of ITL versus number of reference cells}
}
\description{
This is a function that discriminates a target cell population into 2 subgroups, in spatial relationships with the reference cell population. 
In case of lymphocytes and tumor cells, the lymphocytes are discriminated into intra-tumoral lymphocytes (ITL) and adjacent-tumoral lymphocytes (ATL). 
In general, ITL means the subgroup that is in close spatial proximity of the reference cells; ATL means the subgroup that is in distant spatial proximity of the reference cells
}
\author{
Xiao Li, \email{xiao.li.xl2@roche.com}
}
