% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point_pattern_data_uni.R
\name{Point_pattern_data_uni}
\alias{Point_pattern_data_uni}
\title{Main function to generate spatial features using point process data (univariate markers)}
\usage{
Point_pattern_data_uni(
  path = "/Users/lix233/Haystack/5862_cell_centers/",
  file = "f8008cf4-3fe7-4200-a2af-21c27708ae28.csv",
  cell_class_var = "cell_class",
  x_var = "X0",
  y_var = "X1",
  cell_type = "Lymphocyte",
  scale,
  myplot = FALSE
)
}
\arguments{
\item{path}{The path for the directory that contains the data}

\item{file}{The file name in the directory}

\item{cell_class_var}{The column name in each file that indicates the cell class variable}

\item{x_var}{The column name in each file that indicates the cell location x variable}

\item{y_var}{The column name in each file that indicates the cell location y variable}

\item{cell_type}{The cell type one wants to use as the univariate marker}

\item{scale}{The spatial range that user wants to investigate}

\item{myplot}{Whether to plot the results, if available, by default it is set as FALSE}
}
\value{
This function returns the univariate marker features for spatial point pattern data:
\item{g_AUC}{AUC of G-function}
\item{g_r}{the radius at which G-function reaches maximum, i.e. 1}
\item{k_AUC}{AUC of Ripley's K-function}
\item{k_vals_med}{median of K-function at specified range of radius}
\item{k_vals_q1}{1st quantile of K-function at specified range of radius}
\item{k_vals_q3}{3rd quantile of K-function at specified range of radius}
\item{k_vals_max}{maximum of K-function at specified range of radius}
\item{CE}{Clark and Evans Aggregation Index}
}
\description{
This is a function that calculates the features for univariate marker for spatial point pattern data. The methods included are Ripley's K-function, G-function and Clark and Evans nearest neighbor index
}
\author{
Xiao Li, \email{xiao.li.xl2@roche.com}
}
