% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalization_geostat.R
\name{regionalization_geostat}
\alias{regionalization_geostat}
\title{Semi-variogram for geostatistics data}
\usage{
regionalization_geostat(spp_df, from_type, to_type, scale, myplot)
}
\arguments{
\item{spp_df}{The input spatial data.frame, need to have 3 columns: "x_coord", "y_coord" and "cell_class"}

\item{from_type}{The cell type one wants to use as the "from" cell type}

\item{to_type}{The cell type one wants to use as the "to" cell type}

\item{scale}{The spatial range that user wants to investigate}

\item{myplot}{Whether to plot the results, if available, by default it is set as FALSE}
}
\value{
This core function returns the features for semi-variogram, using geostatistics methods
}
\description{
This is the core function that calculates the features for semi-variogram, using geostatistics methods
}
\author{
Xiao Li, \email{xiao.li.xl2@roche.com}
}
