% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalization_lattice.R
\name{regionalization_lattice}
\alias{regionalization_lattice}
\title{The core function to calculate spatial feature using areal data}
\usage{
regionalization_lattice(spp_df, from_type, to_type, scale, myplot)
}
\arguments{
\item{spp_df}{The input spatial data.frame, need to have 3 columns: "x_coord", "y_coord" and "cell_class"}

\item{from_type}{The cell type one wants to use as the "from" cell type}

\item{to_type}{The cell type one wants to use as the "to" cell type}

\item{scale}{The spatial range that user wants to investigate}

\item{myplot}{Whether to plot the results, if available, by default it is set as FALSE}
}
\value{
This core function returns the features for areal data, using spatial lattice process
}
\description{
This is the core function that calculates the features for areal data, using spatial lattice process.
}
\author{
Xiao Li, \email{li.xiao@gene.com}
}
