% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SorensenL.R
\name{SorensenL}
\alias{SorensenL}
\title{A SorensenL function}
\usage{
SorensenL(x, y)
}
\arguments{
\item{x}{The coordinate x}

\item{y}{The coordinate y}
}
\value{
This function returns the Sorensen's L index
}
\description{
This is the function that calculates the Sorensen's L index
}
\author{
Xiao Li, \email{li.xiao@gene.com}
}
