% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_entropy_alltype.R
\name{spat_entropy_alltype}
\alias{spat_entropy_alltype}
\title{Main function to generate spatial entropy features for all cell types involves}
\usage{
spat_entropy_alltype(
  path = "/Users/lix233/Haystack/5813_cell_centers/",
  file = "267a5ed2-b6ee-4973-bf2d-3526b668eda0.csv",
  cell_class_var = "cell_class",
  x_var = "X0",
  y_var = "X1",
  scale = 200,
  side_length = 2000,
  num_FOV = 5,
  set_seed = 42,
  myplot = FALSE
)
}
\arguments{
\item{path}{The path for the directory that contains the data}

\item{file}{The file name in the directory}

\item{cell_class_var}{The column name in each file that indicates the cell class variable}

\item{x_var}{The column name in each file that indicates the cell location x variable}

\item{y_var}{The column name in each file that indicates the cell location y variable}

\item{scale}{The spatial range that user wants to investigate}

\item{side_length}{The side length of squared FOVs}

\item{num_FOV}{Number of FOVs to choose}

\item{set_seed}{set seed for reproducibility}

\item{myplot}{Whether to plot the results, if available, by default it is set as FALSE}
}
\value{
This function returns different version of spatial entropy
\item{shannon}{shannon entropy}
\item{shannonZ_entropy}{co-occurrence based Shannon entropy}
\item{altieri_entropy_SMI}{Spatial mutual information of Altieri entropy}
\item{altieri_entropy_RES}{Global spatial residual of Altieri entropy}
\item{leibovici_entropy}{Leibovici entropy at specified distances}
\item{spat_diversity}{Spatial form of diversity, i.e. Claramunt entropy}
}
\description{
This is a function that calculates the spatial entropy when all cell types are accounted. The methods included are Shannon entropy, co-occurrence based Shannon entropy, Altieri entropy, Leibovici entropy and Claramunt entropy (spatial form of diversity)
}
\author{
Xiao Li, \email{xiao.li.xl2@roche.com}
}
