% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_entropy_onetype.R
\name{spat_entropy_onetype}
\alias{spat_entropy_onetype}
\title{Main function to generate spatial entropy features for one cell type}
\usage{
spat_entropy_onetype(
  path = "/Users/lix233/Haystack/5813_cell_centers/",
  file = "267a5ed2-b6ee-4973-bf2d-3526b668eda0.csv",
  cell_class_var = "cell_class",
  x_var = "X0",
  y_var = "X1",
  cell_type = "Tumor",
  scale = 200,
  myplot = FALSE
)
}
\arguments{
\item{path}{The path for the directory that contains the data}

\item{file}{The file name in the directory}

\item{cell_class_var}{The column name in each file that indicates the cell class variable}

\item{x_var}{The column name in each file that indicates the cell location x variable}

\item{y_var}{The column name in each file that indicates the cell location y variable}

\item{cell_type}{The cell type one wants to use for calculating spatial entropy features}

\item{scale}{The sides of the square grids}

\item{myplot}{Whether to plot the results, if available, by default it is set as FALSE}
}
\value{
This function returns different version of spatial entropy
\item{batty_entropy}{batty entropy}
\item{battyLISA_entropy}{batty's LISA entropy, i.e. Karlstrom and Ceccato's entropy}
}
\description{
This is a function that calculates the spatial entropy when all cell types are accounted. The methods included are batty entropy and battyLISA entropy (Karlstrom and Ceccato's entropy)
}
\author{
Xiao Li, \email{xiao.li.xl2@roche.com}
}
