% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_diversity.R
\name{spatial_diversity}
\alias{spatial_diversity}
\title{Spatial Form of Diversity}
\usage{
spatial_diversity(spp_df, W)
}
\arguments{
\item{spp_df}{The data frame that contains the coordinates and cell class columns}

\item{W}{The study area of the spatial data}
}
\value{
This function returns the spatial form of diversity
}
\description{
This is a function that computes the spatial version of shannon entropy: https://link.springer.com/chapter/10.1007/11556114_14
}
\author{
Xiao Li, \email{xiao.li.xl2@roche.com}
}
