% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Areal_data.R
\name{Areal_data}
\alias{Areal_data}
\title{Main function to generate spatial features using areal data}
\usage{
Areal_data(
  path = "/Users/lix233/Haystack/5862_cell_centers/",
  file = "f8008cf4-3fe7-4200-a2af-21c27708ae28.csv",
  cell_class_var = "cell_class",
  x_var = "X0",
  y_var = "X1",
  from_type = "Lymphocyte",
  to_type = "Tumor",
  scale,
  myplot = FALSE
)
}
\arguments{
\item{path}{The path for the directory that contains the data}

\item{file}{The file name in the directory}

\item{cell_class_var}{The column name in each file that indicates the cell class variable}

\item{x_var}{The column name in each file that indicates the cell location x variable}

\item{y_var}{The column name in each file that indicates the cell location y variable}

\item{from_type}{The cell type one wants to use as the "from" cell type}

\item{to_type}{The cell type one wants to use as the "to" cell type}

\item{scale}{The spatial range that user wants to investigate}

\item{myplot}{Whether to plot the results, if available, by default it is set as FALSE}
}
\value{
This function returns the features for areal data, using spatial lattice process methods
}
\description{
This is a function that calculates the features for lattice data, using spatial lattice process.
Methods included in this function include global and local autocorrelation indices, e.g. Moran's I, Geary's C, Getis-Ord statistics.
Co-localization indices are also calculated, such as Morisita-Horn index, Sorensen index, and Jaccard index
}
\author{
Xiao Li, \email{li.xiao@gene.com}
}
