% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bhattacharyya.R
\name{Bhattacharyya}
\alias{Bhattacharyya}
\title{Bhattacharyya coefficient index}
\usage{
Bhattacharyya(x, y)
}
\arguments{
\item{x}{The coordinate x}

\item{y}{The coordinate y}
}
\value{
This function returns the Bhattacharyya's coefficient index
}
\description{
This is the function that calculates the Bhattacharyya's coefficient index
}
\author{
Xiao Li, \email{xiao.li.xl2@roche.com}
}
