% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geostatistics_data.R
\name{Geostatistics_data}
\alias{Geostatistics_data}
\title{Main function to generate spatial features using geostatistics data}
\usage{
Geostatistics_data(
  path = "/Users/lix233/Haystack/5862_cell_centers/",
  file = "f8008cf4-3fe7-4200-a2af-21c27708ae28.csv",
  cell_class_var = "cell_class",
  x_var = "X0",
  y_var = "X1",
  from_type = "Lymphocyte",
  to_type = "Tumor",
  scale,
  myplot = FALSE
)
}
\arguments{
\item{path}{The path for the directory that contains the data}

\item{file}{The file name in the directory}

\item{cell_class_var}{The column name in each file that indicates the cell class variable}

\item{x_var}{The column name in each file that indicates the cell location x variable}

\item{y_var}{The column name in each file that indicates the cell location y variable}

\item{from_type}{The cell type one wants to use as the "from" cell type}

\item{to_type}{The cell type one wants to use as the "to" cell type}

\item{scale}{The spatial range that user wants to investigate}

\item{myplot}{Whether to plot the results, if available, by default it is set as FALSE}
}
\value{
This function returns the features for point referenced data (denote from_type as type 1 and to_type as type 2), using geostatistics methods
\item{sill_tumor}{ sill of semi-variogram fitting of type 1, assuming Matérn variogram model}
\item{sill_immune}{ sill of semi-variogram fitting of type 2, assuming Matérn variogram model}
\item{range_tumor}{ range of semi-variogram fitting of type 1, assuming Matérn variogram model}
\item{range_immune}{ range of semi-variogram fitting of type 2, assuming Matérn variogram model}
\item{kappa_tumor}{ kappa of semi-variogram fitting of type 1, assuming Matérn variogram model}
\item{kappa_immune}{ kappa of semi-variogram fitting of type 2, assuming Matérn variogram model}
\item{sill_IC_TC}{ sill of cross-variogram fitting from type 2 to type 1, assuming Matérn variogram model}
\item{range_IC_TC}{ range of cross-variogram fitting from type 2 to type 1, assuming Matérn variogram model}
\item{kappa_IC_TC}{ kappa of cross-variogram fitting from type 2 to type 1, assuming Matérn variogram model}
\item{sill_TC_IC}{ sill of cross-variogram fitting from type 1 to type 2, assuming Matérn variogram model}
\item{range_TC_IC}{ range of cross-variogram fitting from type 1 to type 2, assuming Matérn variogram model}
\item{kappa_TC_IC}{ kappa of cross-variogram fitting from type 1 to type 2, assuming Matérn variogram model}
\item{kappa_IK}{ kappa of indicator-variogram fitting of type 1&2, assuming Matérn variogram model}
\item{sill_IK}{ sill of indicator-variogram fitting of type 1&2, assuming Matérn variogram model}
\item{range_IK}{ range of indicator-variogram fitting of type 1&2, assuming Matérn variogram model}
}
\description{
This is a function that calculates the features for point referenced data, using geostatistics methods. 
Semi-variogram and crossvariogram methods are implemented.
}
\author{
Xiao Li, \email{xiao.li.xl2@roche.com}
}
