% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_I_BV.R
\name{moran_I_BV}
\alias{moran_I_BV}
\title{Bivariate Moran's I index}
\usage{
moran_I_BV(x, y, W)
}
\arguments{
\item{x}{The coordinate x}

\item{y}{The coordinate y}

\item{W}{The spatial weight matrix}
}
\value{
This function returns the Bivariate Moran's I feature
}
\description{
This is the function that calculates the Bivariate Moran's I feature
}
\author{
Xiao Li, \email{xiao.li.xl2@roche.com}
}
